cli with click done
