+++ setup.py
@@ -25,7 +25,7 @@ setup(
     package_data={"utensor_cgen": ["templates/*"]},
     entry_points={
         "console_scripts": [
-            "utensor-cli=utensor_cgen.__main__:cli"
+            "utensor-cli=utensor_cgen.cli:cli"
         ]},
     install_requires=[
         'Jinja2',
